database/requirements.txt
# Database Module Requirements
# =============================
# Python package dependencies for XBRL Coordination Database

# Core Database
# -------------
# SQLAlchemy ORM for database operations
sqlalchemy==2.0.23

# PostgreSQL adapter for Python
psycopg2-binary==2.9.9

# Alembic for database migrations (optional, for future use)
alembic==1.13.0

# Configuration Management
# ------------------------
# Environment variable management
python-dotenv==1.0.0

# Utilities
# ---------
# Date/time handling
python-dateutil==2.8.2

# UUID generation (built-in, but listed for clarity)
# uuid - built-in Python module

# Validation
# ----------
# Pydantic for data validation (optional, for future use)
pydantic==2.5.0

# Development Dependencies
# ------------------------
# Testing framework
pytest==7.4.3

# Test coverage
pytest-cov==4.1.0

# Code formatting
black==23.12.0

# Linting
pylint==3.0.3

# Type checking
mypy==1.7.1